Herdening of UI and Persistence-Layer. Also added first tests and extended the TestDataInitializer to create more appropriate testdata containing domain-specific content

+++ src/main/java/de/bbajor/pvs/base/dto/AddressDto.java
@@ -11,7 +11,7 @@ public class AddressDto {
     private Long version;
     private String street;
     private String houseNo;
-    private Double postalCode;
+    private Integer postalCode;
     private String city;
     private String country;
